pyproject.toml
[tool.poetry]
name = "Artaxerxes"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.5"
django = "^2.2"
django-reversion = "*"
django-allauth = "*"
django-konst = "*"
django-airplane = "^1.0.0"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^4.0.0"}
# Use git master for unreleased BCC support, see https://github.com/perenecabuto/django-sendmail-backend/issues/14
django_sendmail_backend = { git = "https://github.com/perenecabuto/django-sendmail-backend", rev = "43d239f" }
poetry = {version = "^1.0.7", optional = true}
mysqlclient = {version = "^1.4.6", optional = true}

[tool.poetry.dev-dependencies]
flake8 = "*"
coverage = "*"
flake8-per-file-ignores = "*"
flake8-bugbear = "*"
flake8-comprehensions = "*"
flake8-mutable = "*"
flake8-mypy = "*"
flake8-docstrings = "*"
flake8-print = "*"
flake8-tuple = "*"
flake8-commas = "*"
flake8-isort = "*"
pyyaml = "*"
django-debug-toolbar = "*"
factory_boy = "^2.12"
bpython = "^0.18"
parameterized = "^0.7.1"
django-hijack = "^2.1.10"
django-hijack-admin = "^2.1.10"
